+++ VTK/Utilities/vtkexodus2/expnam.c
@@ -49,7 +49,7 @@
 *
 * revision history - 
 *
-*  $Id: expnam.c,v 1.1 2006-11-29 18:09:13 dcthomp Exp $
+*  $Id: expnam.c,v 1.2 2008-04-22 21:14:53 david.cole Exp $
 *
 *****************************************************************************/
 
@@ -147,7 +147,7 @@ int ex_put_name (int   exoid,
    start[1] = 0;
    
    count[0] = 1;
-   count[1] = strlen(name) + 1;
+   count[1] = (long)strlen(name) + 1;
    
    if (ncvarput (exoid, varid, start, count, (void*)name) == -1) {
      exerrval = ncerr;
